Upgrade asciify-image from 0.1.5 to 0.1.6 to fix security vulnerabilities
The npm package asciify-image is dependent on two packages minimist
and url-regex that had security vunerabilities. This change updates
their dependencies by running npm install asciify-image@0.1.6 which
was recommended by npm audit.

Url-regex version was vunerable to 'Regular Expression Denial of
Service' which means that an attacker providing a very long string
in String.test can cause a Denial of Service. This vunerability was
critical and its priority was set to high.

Minimist was vunerable to 'Prototype pollution' which means that
arguments are not properly sanitized, allowing an attacker to modify
the prototype of Object, causing the addition or modification of an
existing property that will exist on all objects. This vunerabilitiy
was considered low.

For more informtion regarding the security vulnerabilities:

https://snyk.io/vuln/SNYK-JS-MINIMIST-559764
https://snyk.io/vuln/SNYK-JS-URLREGEX-569472

Minimist: https://www.npmjs.com/advisories/1179
Url-regex: https://www.npmjs.com/advisories/1550